Seperate ListBoxTemplate XML data to string and class (wip) Adjust Animation and style
