+++ Flow.Launcher/Languages/en.xaml
@@ -196,4 +196,8 @@
     <system:String x:Key="update_flowlauncher_update_files">Update files</system:String>
     <system:String x:Key="update_flowlauncher_update_upadte_description">Update description</system:String>
 
+    <!--General Guide & Hotkey-->
+    <system:String x:Key="SmallTip01">Context Menu : &#129046; or Shift+Enter | Containing Folder : Ctrl+Enter</system:String>
+    <system:String x:Key="SmallTip02">Back to Result : &#129044; or ESC</system:String>
+    
 </ResourceDictionary>
